requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
asyncpg==0.30.0
bcrypt==4.2.0
python-jose[cryptography]==3.3.0
pydantic[email]==2.9.0
python-dotenv==1.0.1